Bump qdrant client version, add eta handling, fix reflection issues, add file checks

+++ src/main/java/io/metaloom/qdrant/cli/command/action/impl/CollectionAction.java
@@ -1,5 +1,6 @@
 package io.metaloom.qdrant.cli.command.action.impl;
 
+import static io.metaloom.qdrant.cli.ExitCode.FILE_ERROR;
 import static io.metaloom.qdrant.cli.ExitCode.OK;
 import static io.metaloom.qdrant.cli.ExitCode.SERVER_FAILURE;
 
@@ -16,8 +17,11 @@ import org.slf4j.LoggerFactory;
 import io.metaloom.qdrant.cli.ExitCode;
 import io.metaloom.qdrant.cli.command.QDrantCommand;
 import io.metaloom.qdrant.cli.command.action.AbstractAction;
+import io.metaloom.qdrant.client.http.QDrantHttpClient;
+import io.metaloom.qdrant.client.http.impl.HttpErrorException;
 import io.metaloom.qdrant.client.http.model.collection.CollectionDescription;
 import io.metaloom.qdrant.client.http.model.collection.CollectionListResponse;
+import io.metaloom.qdrant.client.http.model.collection.CollectionResponse;
 import io.metaloom.qdrant.client.json.Json;
 
 public class CollectionAction extends AbstractAction {
@@ -34,19 +38,21 @@ public class CollectionAction extends AbstractAction {
 			if (isSuccess(response)) {
 				List<CollectionDescription> collections = response.getResult().getCollections();
 
+				long totalWritten = 0;
 				if (outputPath.equals("-")) {
 					try (BufferedWriter writer = new BufferedWriter(new OutputStreamWriter(System.out))) {
-						dump(collections, writer);
+						totalWritten = dump(client, collections, writer);
 					}
 				} else {
 					File outputFile = new File(outputPath);
-					// if (!outputFile.canWrite()) {
-					// log.error("Unable to write to " + outputFile.getAbsolutePath());
-					// System.exit(12);
-					// }
-					try (BufferedWriter writer = new BufferedWriter(new FileWriter(outputFile))) {
-						dump(collections, writer);
+					if (outputFile.exists()) {
+						log.error("Output file already inplace.");
+						return FILE_ERROR;
 					}
+					try (BufferedWriter writer = new BufferedWriter(new FileWriter(outputFile))) {
+						totalWritten = dump(client, collections, writer);
+					}
+					log.info("Backup of {} collections written to {}", totalWritten, outputPath);
 				}
 
 				return OK;
@@ -57,11 +63,27 @@ public class CollectionAction extends AbstractAction {
 		});
 	}
 
-	private void dump(List<CollectionDescription> collections, BufferedWriter writer) throws IOException {
+	private long dump(QDrantHttpClient client, List<CollectionDescription> collections, BufferedWriter writer)
+			throws IOException, HttpErrorException {
+		long totalWritten = 0;
 		for (CollectionDescription collection : collections) {
-			String json = Json.parseCompact(collection);
-			writer.append(json + "\n");
+			String collectionName = collection.getName();
+			if (log.isDebugEnabled()) {
+				log.debug("Loading collection {}", collectionName);
+			}
+			CollectionResponse response = client.loadCollection(collectionName).sync();
+			if (isSuccess(response)) {
+				String json = Json.parseCompact(response.getResult());
+				writer.append(json + "\n");
+				totalWritten++;
+			} else {
+				log.error("Error while loading info for collection " + collectionName + ". Got status: "
+						+ response.getStatus());
+				throw new RuntimeException("Error while loading collection " + collectionName);
+			}
+
 		}
+		return totalWritten;
 	}
 
 }
